Assignment_01/pyproject.toml
[project]
name = "product_suggester_ai_agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.2.1",
    "python-dotenv>=1.1.1",
]
